model-service/pyproject.toml
[tool.poetry]
name = "model-service"
version = "0.1.0"
description = "Text-to-SQL service using Ollama"
authors = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
httpx = "^0.27.0"
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
pyyaml = ">=6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
anyio = {extras = ["trio"], version = "^4.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
